+++ modules-core/prov-model/src/main/java/module-info.java
@@ -6,6 +6,7 @@ module prov.model {
     requires org.apache.commons.io;
     requires org.apache.commons.lang3;
     requires org.apache.logging.log4j;
+    requires com.fasterxml.jackson.annotation;
 
     exports org.openprovenance.prov.model;
     exports org.openprovenance.prov.vanilla;
@@ -22,6 +23,7 @@ module prov.model {
     opens org.openprovenance.prov.vanilla;
     opens org.openprovenance.prov.model;
     opens org.openprovenance.prov.model.interop;
+    opens org.openprovenance.prov.rules;