+++ modules-core/prov-model/src/main/java/org/openprovenance/prov/model/interop/ApiUriFragments.java
@@ -4,6 +4,7 @@ public interface ApiUriFragments {
     String FRAGMENT_DOCUMENTS_FORM = "/documents_form/";
     String FRAGMENT_DOCUMENTS = "/documents/";
     String FRAGMENT_METRICS = "/metrics/";
+    String FRAGMENT_PROVTYPEMAP = "/typemap";
 
     String FRAGMENT_PROVAPI = "/provapi";
     String FRAGMENT_RESOURCES = "/resources/";