+++ modules-services-metrics/service-metrics/src/main/java/org/openprovenance/prov/service/metrics/TmpTypeMap.java
@@ -1,0 +1,7 @@
+package org.openprovenance.prov.service.metrics;
+
+public class TmpTypeMap {
+    public String map;
+    public String set;
+    public String list;
+}