+++ modules-services-metrics/service-metrics/src/main/java/org/openprovenance/prov/service/metrics/inface/BiConsumerWithException.java
@@ -1,0 +1,16 @@
+package org.openprovenance.prov.service.metrics.inface;
+
+import java.sql.SQLException;
+
+@FunctionalInterface
+public interface BiConsumerWithException<T, U> {
+
+    /**
+     * Applies this function to the given arguments.
+     *
+     * @param t the first function argument
+     * @param u the second function argument
+     * @throws SQLException a SQL exception
+     */
+    void accept(T t, U u) throws SQLException;
+}