+++ modules-services/prov-service-core/src/main/java/org/openprovenance/prov/service/core/SwaggerTags.java
@@ -4,6 +4,7 @@ public interface SwaggerTags {
     String VALIDATION="validation";
     String TEMPLATE="template";
     String DOCUMENTS="documents";
+    String TYPES="types";
     String RANDOM="random";
     String VIEW="view";
     String RESOURCES ="resources";