+++ modules-summarisation/prov-typemap/src/main/java/org/openprovenance/prov/java/typemap/TypeMapProcessor.java
@@ -1,0 +1,5 @@
+package org.openprovenance.prov.java.typemap;
+
+public interface TypeMapProcessor<T> {
+    T process(String map, String set, String list);
+}